Renamed to FluentStoreApiClient
